requirements.txt
# hmac
pyyaml